修复所有DEV1中的问题
