--- example/.env.example
@@ -1,11 +1,0 @@
-ENVIRONMENT=dev
-DRIVER=~fastapi
-PORT=8080
-HOST=127.0.0.1
-LOCALSTORE_USE_CWD=true
-LOG_DIR=logs
-BOT_NAME=Amrita
-RATE_LIMIT=5
-WEBUI_ENABLE=true
-WEBUI_USER_NAME=admin
-WEBUI_PASSWORD=admin123
