--- example/templates_example/my_page.html.jinja2
@@ -1,16 +1,0 @@
-{%extends base_html%}
-{% block title %}我的Amrita页面{% endblock %}
-{% block topbar_title %}我的页面{% endblock %}
-{% block content %}
-<div class="container">
-    <div class="row">
-        <div class="col-md-12">
-            <div class="panel panel-default">
-                <div class="panel-heading">
-                    <h3 class="panel-title">我的页面</h3>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
