+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "amrita"
-version = "1.0.0.dev1"
+version = "1.0.0.dev2"
 description = "A powerful AI bot framework powered by NoneBot2"
 readme = "README.md"
 license = "GPL-3.0-or-later"
@@ -30,7 +30,7 @@ dependencies = [
 
 [project.optional-dependencies]
 full = [
-    "amrita-core==0.4.2.1",
+    "amrita-core==0.4.3.2",
     "nonebot-plugin-uniconf>=0.1.3",
     "aiomysql>=0.3.2",
     "aiopg>=1.4.0",
